Reuse the layout state object.
livesplit-core recently added a feature to reuse a layout state object.
In theory this should be slightly faster because less memory allocations
need to be done but in practice I haven't seen much of a difference. I'm
still going to do this anyway as its pretty easy to implement and any
performance improvement is nice.